backend/requirements.txt
Django==4.1.4
django-rest-framework==0.1.0
